Universal login state
Three new things about the Firebase 3 SDK:

* It no longer allows us to choose session-only login persistence. Once
  you’re logged in, you’re logged in.
* Login sessions are browser-wide. If you’re logged in in one tab,
  you’re logged in in all tabs. Changing login state in one tab will
  change your login state in all tabs.
* It doesn’t provide authentication credentials (e.g. GitHub access
  token) when you ask for current login state or auth state change. This
  means that there is no way to get credentials when we re-authenticate
  when the environment loads.

Because Firebase sessions are universal, we have to make Popcode login
state universal as well. To that end, logging in in one tab logs you in
in all tabs. Logging out of one tab logs you out of all tabs.

The first step here is to start storing the GitHub credentials in the
user’s own Firebase data, rather than in the browser session. This
means that there is no situation in which the user can be logged in to
Firebase but without an available GitHub token. So, we can guarantee
that application login state is consistent with Firebase session state.

We refactor the `firebaseAuth` module to use a reactive Observable to
monitor auth state change; this Observable powers the initial
auth state promise that is used for bootstrap, as well as a pair of
observer hooks that can be used to detect subsequent login state change.

If the user signs in or signs out of Firebase in a different tab, the
change is propagated to all Firebase sessions in the browser. This means
that the application should update its sign in state accordingly.
Previously, we only updated login state after an explicit user action to
log in or log out. Now, instead, we listen for login state changes and
update application state accordingly.