Session expiry
We do not want students to be logged in to Popcode indefinitely, but
Firebase 3 only supports long-lived sessions. So, we store a cookie that
holds the UID of the currently logged in user. This cookie expires after
five minutes, but it is refreshed every second that Popcode is running.
The effect is that the student stays logged in as long as they have any
Popcode tab open, and for five minutes thereafter. After that, the
Popcode session expires and they must log in again.