Keep project when logging out if locally modified
The following scenario is fairly common:

* Student logs in to Popcode
* Student does some work
* Student then realizes they are logged in to the wrong account (because
  the browser was logged in to a different GitHub)
* Student “loses” their work because logging out resets the workspace

While the work the student did was, in fact, still available in the
account they were logged in to, they may not have access to it, and this
is generally an inconvenient situation.

Thus, as part two of the invariant that changing login state does not
change the current project, we retain the current project when logging
out.