Use Immutable root state for Redux store
To date we have used the built-in `combineReducers` from Redux to put
together the output of different reducers into the single top-level
store state. However, this creates an awkward shape for the state,
namely that the top-level state is a plain JavaScript object, and all
of its properties are Immutable objects.

Instead, use `redux-immutable`, which provides an implementation of
`combineReducers` that merges subtrees into an Immutable Map, rather
than a plain object. This creates a much more consistent situation where
the entire state tree is Immutable.

This has been bothering me for a while.