Use providerData for profile data if available
Apparently the `displayName` and `photoURL` methods are not consistently
available on the root Firebase user object. These properties are still,
however, available on the `providerData`, which contains raw information
from GitHub. So, prefer the first `providerData` record available as the
source of profile data, falling back to the root object.