More descriptive error when auth fails because of network error
Ideally, we’d like to retry sign in if a network error causes it to
fail. However, our only interface to signing in is to call the
`signInWithPopup()` method of Firebase, which results in a blocked popup
if it’s not called in an execution context that was initiated by a user
action.

So, just fail with a slightly more descriptive error, and don’t send the
error on to Bugsnag.