You can’t be missing a semicolon from before the first line
The `prettycss` validator tries to heuristically determine whether an
`extra-tokens-after-value` parse error happens because of actual
superfluous tokens, or because a semicolon is missing from the previous
line. However, there was no check to see if the “previous line” actually
exists, meaning superfluous tokens on the first line will be detected as
a missing semicolon on the nonexistent line before it. Add a check to
make sure that does not happen.