Test directly for whether error token is the first token
Previously, we were using lodash `startsWith` to see if the token
content reported by the PrettyCSS `extra-tokens-after-value` error
matched the beginning of the line in question. However, this was a
simple substring match, meaning that the error token might be a *prefix*
of the first token on the line.

This means that in an edge case like:

```css
button{border:20px solid b;
    }
```

The error token at the end of the line, `b`, would match the beginning
of the line, since it begins with `b`.

Instead, use the error token’s character index on the line to directly
test whether it is the first token on the line, by checking to see if it
is the index of the first non-whitespace character on the line.