Don’t transpile import statements
Webpack handles ES6 import statements natively, and is able to do
tree-shaking if we use them. So, disable module transpilation.

This does mean that our babel gulpfile can’t use ES6 imports, because it
runs in node. That’s the only actual point of using gulpfile.babel.js,
so rename it normal gulpfile.js and switch to using CommonJS.
Unfortunately we also have to do that for src/config.js, which is shared
between the gulpfile and the application code.