async/await
Now that we’re officially ES2017, use async/await instead of explicit
`.then()` when handling the outcome of a promise.

Generally makes the code easier to work with, although I don’t love
going back to `try`/`catch` and actually find the two-callback approach
to `then()` better—in particular because it makes it natural to handle
errors that occurred when executing the promise, without also
inadvertently catching errors in one’s own handler for the promise.