First saga
Initial integration of `redux-saga`, which I hope to use to completely
replace `redux-thunk` for coordination of complex or asynchronous
actions.

The first saga is `createProject`, which consumes the `CREATE_PROJECT`
action, generates a new project key, and dispatches the
`PROJECT_CREATED` action with that key.

This commit also integrates the redux-saga middleware into the redux
store.