Add errors saga
Errors saga handles validation. Currently listens for
`CHANGE_CURRENT_PROJECT` and validates all sources when this action is
dispatched.

This is notably different from the old approach taken by the
`changeCurrentProject` thunk action, which procedurally coordinated both
the saving of the project and the initiation of validation in one place.
Sagas allow us to be more reactive about it, with different concerns
initiating different asynchronous operations in response to the same
action.