Loop-breaker take 2
Restores the [switch to
loop-breaker](https://github.com/popcodeorg/popcode/pull/726), which I
had to revert because certain JavaScript inputs were causing the
environment to throw an uncaught exception.

It turns out that the main problem was code that had a single slash at
the beginning of the line (this was likely intended to be a double-slash
for a comment). This causes esprima to throw an “invalid regular
expression” error.

In order to fix the problem, modify the esprima validator to *always*
return an error if the source did not successfully parse. Unlike with a
linter, we can safely assume that any error thrown by esprima should be
reported as a validation error, however vague, by Popcode.