Absolutely position column contents
If a very long token (e.g. a data URI) appears in an error message, the
column will expand to fit its contents, causing it to consume the entire
width of the container, and causing the editors to disappear.

I couldn’t find a completely satisfactory solution using flex all the
way down; the interaction between the contents size and the flex
properties is not well-documented, but it is certainly sensitive to the
size of the content.

By wrapping the contents of each column in an absolutely positioned div
that fills its container horizontally and vertically, we can get exactly
the behavior we want.  The use of absolute positioning here means that
the contents of the columns cannot affect the layout of the columns; it
breaks them out of the flow. Instead, the columns’ width is determined
solely by the flex arrangement, and the contents are forced to stay
within the confines of the columns.

With this in place, a simple `overflow-wrap: break-word` gives us the
exact wrapping behavior we want for the long tokens.