importGist saga takes an `action` argument
Best to be consistent and say that, if a saga takes any arguments, the
final argument should be an `action`. This allows any saga to be used as
a handler for a `takeEvery`; usually it should be sufficient for the
saga that initially handles an action to pass that action on to any
delegate sagas.