Remove redundant logIn action creator
`logIn` is a thunk-style action creator that handles user login as well
as the asynchronous operations that go along with it. Happily, the
existing reducer and saga handlers already do the right thing when
`USER_AUTHENTICATED` is dispatched, because they need to for application
bootstrap to work properly. So, simply remove `logIn`.