Error reducer consumes updateProjectSource and initiates validation
First time using the `redux-saga` task validation. In particular, if a
user is typing quickly, we want to make sure to only pay attention to
the results of the validation for the last thing they typed in.

More generally, we only want to pay attention to the *last* validation
initiated for a given language. To that end, `validateSource()` is given
a map of languages to the currently ongoing validation task for that
language, if any. This map is maintained inside the root saga and passed
to all sagas in the module. If, when `validateSource()` is called, there
is a concurrent validation task for the same language, it will cancel
that task. When `validateSource()` completes, it clears its own
validation task from the map.