Workspace watches for change to `gists.currentExport.status`
If it’s ready, attempts to display the gist export, then dispatches
the appropriate action indicating that the gist was either displayed or
it wasn’t.

If the status has changed to errored, clean up after the gist export
by closing the tab and removing the `window` instance from instance
variable data.