Reducer unhides editor when line focused
Previously the Workspace component’s handler for line focus would
dispatch an `unhideComponent` action from the handler for a focused line
request. This had trivially buggy behavior in that it omitted the
required first argument to `unhideComponent`, which should have been the
current project ID.

However, it also doesn’t follow the recently preferred pattern of
reducers knowing what to do in response to actions, rather than
procedural code dispatching multiple actions to deal with different
aspects of responding to a particular user action or external event.

So, remove the call to `unhideComponent` and instead have the `ui`
reducer handle the focus line action.