Use babel-preset-env rather than babel-preset-latest
`babel-preset-latest` has been deprecated for awhile. Instead, use
`babel-preset-env`.

Because we want to use the same source of truth for “what browsers do we
support?”, we remove the static `.babelrc` configuration file and
instead construct the Babel configuration in `webpack.config.js`. This
is fine because Babel is exclusively used through the Webpack
`babel-loader`.

The configuration respects the browser list in `config/browsers.json`,
which is also used by `autoprefixer` and by the actual application to
prevent users with unsupported browsers from attempting to run the app.

Also introduces a `DEBUG` environment variable, which if set to true
instructs Babel to compile for the latest Chrome version (i.e. it
doesn’t need to do much if anything). This is intended to make debugging
easier, since it’s often pretty difficult to step through ES2017
compiled down to ES5, especially if there is a generator or async/await
involved.