Re-add no-unused-prop-types linter
The linter’s logic has gotten better; we needed to change a couple of
things around in the code, but nothing that has made the code
particularly worse. In places where I was doing anything with props, I
also refactored to destructure.

This linter will be especially useful as we extract out container
components, to help catch props in the old monolithic hierarchy that are
no longer needed.