Combine errors and runtimeErrors
Having a separate `runtimeErrors` key at the top level of the store
always felt bolted on. In order to simplify the logic in the
`ErrorReport` component, undertake to fix that bolting on.

Runtime errors are now stored in the normal `errors` store. The
distinction is made in the `state` key within each language’s errors. If
an error was generated during validation, the state is set to
`validation-error`. If it was generated at runtime, the state is set to
`runtime-error`. These two states replaces the previous `failed` state.

The various components are thus able to figure out what to do based on
the error states. In particular, if the “dominant” error state is
`runtime-error`, the error list will be shown in the “docked” variant.

This does add a minor amount of complexity in some  places where
components are checking the error state, where now they need to look for
two states rather than one. Tellingly, though, they aren’t always
looking for the _same_ two states, suggesting that this is a useful
distinction to make.