Add ProjectPreview container
Adds a container for the `<ProjectPreview>` component. This uses the
higher-order form of `mapStateToProps()`, which returns a function that
returns a separate mapping function for each instance of the component.
Each mapping function has its own `makeGetProjectPreview` that maintains
its own single-element cache of previous results. The effect is that we
are able to memoize all of the project previews we have generated, and
the memo is naturally garbage collected when the preview component
ceases to exist.

I don’t think this is totally the end of the story for the “Load
Project” performance problems. Initially opening the list with a lot of
projects contained therein is still going to be slow. I think we can
probably make the actual code that generates the preview more efficient,
and maybe can also experiment with lazy rendering.