Serve all static assets from the same root directory
Previously assets were served from the `static/` directory; at the root
of this directory were unmodified static files that were part of version
control, e.g. `index.html`. This contained `static/compiled/`, which
contained compiled files (CSS bundle, JS bundle, etc.) and was not under
version control.

Now all assets are served from the `dist` directory in the root of the
project; `static` has moved to `src/static`.

The proximate motivation for this is that service workers can’t control
files outside of the directory they are served from; so, serving a
service worker from `/compiled/sw.js` prevented that service worker from
caching `/index.html`. This can be overridden with a header, but since
we serve the site statically from S3, that’s not so simple.

This change also I think is strictly more sensible.