Display a checkbox after snapshot link is copied
We want to give some sort of feedback when the snapshot link is
successfully copied. So, just display a checkbox next to the message
once it’s been copied.

Actually rather complicated! We needed to add a new type of state to the
UI store, representing arbitrary metadata associated with a particular
notification. Notifications can set their own metadata and then have
access to it.

This seems to be a specific case of the general idea of “UI” or “local”
state—state that is only meaningful to a specific component, which the
rest of the application would never need to know about. There are
various libraries that endeavor to solve this problem directly; however,
none of them appear perfectly suited to the task at hand. If we need
local state more extensively, I’d be tempted to hand-roll our own
higher-order component to get that.

Also creates a `NotificationList` container component, which was on my
list anyway.