Flex rows at top level of layout
Previously the top-level layout was based on a flex row, containing the
dashboard, sidebar, and main editing environment. This worked pretty
well because usually we didn’t have anything else that needed horizontal
space. The exception was notifications, which ended up just pushing the
rest of the environment down a bit, such that the bottom was outside the
viewport—definitely a bug, but not the end of the world.

This fixes that bug by creating a top-level column flex and then within
it a row flex. The column flex currently just contains any notifications
followed by the main user interface. However, the real motivation for
this is to allow us to add an always-on  top bar for the layout
overhaul, which will follow in a subsequent PR.