Store raw profile data and return when logging back in
The Firebase `UserCredential` object can optionally contain an
`additionalUserInfo` key that contains the raw profile data from the
provider. This is useful e.g. in the case of GitHub (the only case!)
for retrieving the user’s username (which, frustratingly, is not
available via the normalized Firebase user profile).

However, when re-authenticating (e.g. loading Popcode while already
logged in), that data is not made available to the auth event handler,
and as far as I can tell there is no way to ask for it. So, whenever the
user logs in explicitly, we grab the `additionalUserInfo` and store it
in the Firebase database; then we retrieve it on re-authentication and
add it to the synthetic `UserCredential` object that is exposed to
reauthentication listeners. This is exactly the same thing we do for the
`credential` part of the `UserCredential` already.