Don’t fully compile project just to extract title
Previously the preview generator ran the full preview, which includes
making substantial modifications to the project’s DOM, for each project
we wanted to display in the project picker. But the only thing we
actually do with that information is pull the title from the document!
Oy.

Now we just have a separate code path that parses the document and
returns its title. This makes the project picker MUCH faster when you
have a substantial number of projects.