Component factory for top bar menus
Builds a component factory for top bar menus, which looks vaguely like a
higher-order component but doesn’t really fit the pattern. Specifically:

* The `createMenu()` function takes an object with properties `name` and
  `mapPropsToItems`, which returns a function…
* That takes two component classes, a `Label` class and an `Item` class.

This will then return a highly decorated component that renders the
menu, deals with opening/closing, marking enabled items as enabled, etc.

One of the decorations is the `onClickOutside` HOC which properly
handles clicks outside the menu (closing it).