Give menus more control over how items are rendered
Initial implementation had `createMenu` take a `mapPropsToItems`, which
returned an array of objects containing properties `key`, `isEnabled`,
and `props`. The higher-order component was then passed a component
class which actually rendered items.

As well as setting up a structure that’s both rigid and not particularly
discoverable, this meant that menus were assumed to consist of a
homogeneous collection of items, each of which is rendered in the same
way and behaves in the same way when clicked. This is not true, however,
of things like the hamburger menu.

So, instead, `createMenu` exports a `<MenuItem>` component, which is
used to wrap the actual contents of a menu item and takes `key`,
`isEnabled`, and `onClick` props. The factory now takes a `renderItems`
function, which can return any valid JSX node. This also improves the
interface of generated menu components, since the click handler can now
be called whatever feels natural.