Smooth preview refresh
Typing in the HTML is unpleasant for two reasons:

1. A spinner briefly appears while the project is asynchronously
   validating. This is a fairly recent change, with the introduction of
   custom validation, but it’s not a problem with the validation itself.
   Validation is expected to be asynchronous and can take a non-trivial
   amount of time. The UX around it is bad.
2. Even before the spinner issue, there was a noticeable flash as the
   preview iframe refreshed itself.

This is a rough-and-ready sketch of a fix for these issues. Instead of
using a single iframe and updating its `srcdoc` property, we instead
create a new frame every time the project changes. We also keep the
frame from the previous version of the compiled project in the DOM,
layering the two on top of each other with absolute positioning.
Further, we set the new frame’s visibility to hidden until it has
loaded. The effect is that, for the most part, one can’t see any flashes
because each new frame instance appears fully-loaded on top of the last.

There is a bunch of work to do around the behavior of error messages,
which are basically broken at the moment.