Default all htmllint rules to false
Our troubles with `attr-validate` began when it was introduced as a new
rule and on by default. Unfortunately `htmllint` doesn’t provide
out-of-the-box functionality to disable all rules, so we just iterate
over `htmllint.rules` and set anything we haven’t explicitly enabled to
`false`. Future-proofs us against inadvertently enabling new rules when
upgrading.