Group Bugsnag errors by message when stacktrace includes minified files
E.g. `firebase/auth` is minified and does not have a sourcemap, so the
file numbers in stacktraces that include `firebase/auth` are meaningless
and change with each release. So, for stacktraces that include stack
frames in `firebase/auth`, we want to ignore the stacktrace and just
group based on the error. This will allow us to more effectively ignore
a couple of errors that come from `firebase/auth`, generate a lot of
noise, but do not seem to have any user-facing effect.