Auth channel
Consolidates all login/logout business logic into the `user` saga, where
it belongs.

First, uses a Redux Saga channel to represent the login/logout state,
which we can then just watch with a saga, which allows us to do the same
thing we were doing with `rxjs`, but in a way that feels much more
natural in this code base. Allows us to remove the code in the
`Workspace` component (of all places) that monitored Firebase
login/logout state and dispatched actions accordingly.

Also creates `LOG_IN` AND `LOG_OUT` actions that are consumed by the
`user` saga, which makes the appropriate calls to the Firebase client
and handles errors where appropriate. Bonus: this behavior is now under
test!