Explicitly set preview srcdoc after attaching load handler
In production the `load` handler is never firing; it’s hard to tell why
for sure, but one possibility is that the frame is loading before we get
the `ref` and attach the handler. So, always set the `src` initially to
`about:blank`, and then in the `ref` callback set the `srcdoc` to the
actual preview contents *after* attaching the load listener.

Should guarantee that the load listener fires.