Initial Redux-level support for console
Sets up a `console` section in the Redux hierarchy that holds console
state. This is simply an `OrderedMap` collection, where each element is
a `Record` which represents a single read-eval-print iteration in the
console. The records contain the expression, a `state`, and an optional
`value` (which is only meaningful if the `state` is `"evaluated"`).

The only action so far is the one that handles a new line being input
into the console.