Full REPL flow in console
`<PreviewFrame>` component now handles console value messages, and
propagates them through a new action that updates the appropriate input
with a result. Errors are propagated in a similar fashion. `<Console>`
component properly displays the output.

Also renames the `ConsoleInput` record/entity to `ConsoleEntry` which
seems more descriptive.

Finally, removes the concept of an entry’s status. It’s not really
useful—mostly we just care if there is a value or an error attached to
the entry.