Upgrade to jQuery 3.2.1
Upgrades from jQuery 2.2.3 to jQuery 3.2.1. This is a major version
upgrade with [numerous breaking
changes](https://jquery.com/upgrade-guide/3.0/), but none of them look
like they would have any effect on the kind of jQuery students write in
Popcode. In fact, the only thing that even comes close is the behavior
of `show()`, which is in fact the proximate motivation for this upgrade:
calling `show()` on an element that was hidden with a CSS tag selector
inside a sandboxed iframe threw an error in jQuery 2. Now it just sets
its `display` value to `block` (regardless of whether this is the
correct `display` for the element in question).