Refresh preview just adds copy of compiled project to state
Previously we refreshed the preview by artificially adding the idea of a
“last refresh timestamp”, which was given as an argument to the preview
generator, causing its output to change and thus the preview iframe to
refresh.

Things work differently now.

Since the preview now has first-class state, all we need to do is add a
copy of the current compiled project to the stack of compiled projects,
which triggers the desired refresh behavior in the Preview component.