Hide and show console
Adds console as a UI component that can be hidden and shown. This is the
first component that is hidden by default, which makes dealing with
existing data a bit awkward—how do we tell the difference between a
project that has explicitly opened the console, and one that was saved
before the console became a thing?

For now, just punt and always hide the console when rehydrating a
project (which is perhaps perfectly fine behavior anyway).