Don’t assume window properties on preview frame are defaults
When we reference `window.name`, `window.parent`, etc., we’re assuming
those are the browser default globals. However, those can be overwritten
by both project JavaScript code and expressions entered into the
console. So, grab references to the values we’re interested in right
when the page loads, so we’re not tripped up if they are later
redefined.