Load preview support code as Webpack bundle
The code that runs in the preview frame is getting ever more
substantial. So, let’s make it a Webpack bundle instead of a series of
hacky and fragile function-to-strings.

In order to figure out the correct URL for the preview bundle (including
content hash in production), we load it onto the page using
`HtmlWebpackPlugin`, but add in the ScriptExt extension which we use to
give the script tag a nonsensical `type` attribute, meaning it doesn’t
load. We can then, in the main bundle, find the script tag, read its
`src` attribute, and load the JavaScript using an AJAX request.

Right now the preview bundle is just one big module, but we’ll improve
that.